+++ CHANGELOG.md
@@ -3,6 +3,24 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [0.0.2] - 2018-01-17
+### Added
+- RemoveButton Component.
+- ProgressBar Component.
+- Updated readme to reflect the new RemoveButton and ProgressBar Component.
+
+### Changed
+- Moved remove input and ProgressBar out of Add & AddMedia into there own components.
+- Replaced code of date of measurements with InputGroup Component.
+
+### Fixed
+- Duplicate code in AddMedia.
+- Reduced lines of code in the AddMedia Component.
+
+### Removed
+- Removed UID prop from Add component as it doesn't use / need it.
+- Removed trailing unused code.
+
 ## [0.0.1] - 2018-01-17
 ### Added
 - InputGroup Component now accepts the props fullWidth & placeholder.